+++ billinghelper/src/main/kotlin/com/vojtkovszky/billinghelper/BillingListener.kt
@@ -6,7 +6,13 @@ interface BillingListener {
     /**
      * @param event one of resolved [BillingEvent]s
      * @param message as reported by [BillingResult.getDebugMessage]
-     * @param responseCode as reported by [BillingResult.getResponseCode] as [BillingClient.BillingResponseCode].
+     * @param responseCode as reported by [BillingResult.getResponseCode] as [BillingClient.BillingResponseCode]
+     * @param subResponseCode as reported by [BillingResult.getOnPurchasesUpdatedSubResponseCode] as [BillingClient.OnPurchasesUpdatedSubResponseCode]
      */
-    fun onBillingEvent(event: BillingEvent, message: String?, responseCode: Int?)
+    fun onBillingEvent(
+        event: BillingEvent,
+        message: String?,
+        @BillingClient.BillingResponseCode responseCode: Int?,
+        @BillingClient.OnPurchasesUpdatedSubResponseCode subResponseCode: Int?
+    )
 }